requirements.txt
langchain_openai
langchain_core
langchain_community
python-dotenv
streamlit
langserve
ipykernel
bs4
pypdf
chromadb
faiss-cpu
groq
cassio
beautifulsoup4
langchainhub
PyPDF2
langchain-groq